+++ src/MonadIO/Process.hs
@@ -1,5 +1,5 @@
 module MonadIO.Process
- ( doProc, system, systemx )
+ ( doProc, system, systemx, systemN, systemS )
 where
 
 import Prelude  ( (-), fromIntegral )
@@ -29,7 +29,6 @@ import Data.Textual  ( Printable )
 
 -- fpath -------------------------------
 
-import FPath.AbsFile           ( AbsFile )
 import FPath.Error.FPathError  ( AsFPathError )
 
 -- monaderror-io -----------------------
@@ -43,11 +42,10 @@ import Data.MoreUnicode.Either  ( pattern 𝕷, pattern 𝕽 )
 import Data.MoreUnicode.Lens    ( (⊣) )
 import Data.MoreUnicode.Maybe   ( pattern 𝕹 )
 import Data.MoreUnicode.Monad   ( (≫) )
-import Data.MoreUnicode.Text    ( 𝕋 )
 
 -- mtl ---------------------------------
 
-import Control.Monad.Except  ( ExceptT, MonadError, throwError )
+import Control.Monad.Except  ( MonadError, throwError )
 
 -- process -----------------------------
 
@@ -59,8 +57,9 @@ import System.Process  ( ProcessHandle, waitForProcess )
 
 import MonadIO.Error.CreateProcError  ( AsCreateProcError )
 import MonadIO.Error.ProcExitError    ( AsProcExitError, asProcExitError )
-import MonadIO.Process.CmdSpec        ( CmdSpec
-                                      , anyExit, cmdSpec, expExit, mkCmd )
+import MonadIO.File                   ( devnull )
+import MonadIO.NamedHandle            ( stdin )
+import MonadIO.Process.CmdSpec        ( CmdSpec, expExit )
 import MonadIO.Process.ExitStatus     ( ExitStatus( ExitSig, ExitVal ) )
 import MonadIO.Process.MakeProc       ( MakeProc, makeProc )
 import MonadIO.Process.MkInputStream  ( MkInputStream )
@@ -135,6 +134,28 @@ system inh cspec = do
   then throwError $ asProcExitError cspec exit (toMaybeTexts w)
   else return (exit,w)
 
+--------------------
+
+{- | Like `system`, but implicitly takes `/dev/null` for input. -}
+systemN ∷ ∀ ε ζ ω μ .
+          (MonadIO μ, ToMaybeTexts ω,
+           MakeProc ζ, OutputHandles ζ ω, HasCallStack,
+           AsCreateProcError ε, AsFPathError ε, AsIOError ε, AsProcExitError ε,
+           MonadError ε μ, HasCallStack) ⇒
+          CmdSpec → μ (ExitStatus, ω)
+systemN c = devnull ≫ \ l → system l c
+
+--------------------
+
+{- | Like `system`, but implicitly takes `stdin` for input. -}
+systemS ∷ ∀ ε ζ ω μ .
+          (MonadIO μ, ToMaybeTexts ω,
+           MakeProc ζ, OutputHandles ζ ω, HasCallStack,
+           AsCreateProcError ε, AsFPathError ε, AsIOError ε, AsProcExitError ε,
+           MonadError ε μ, HasCallStack) ⇒
+          CmdSpec → μ (ExitStatus, ω)
+systemS c = system stdin c
+
 ----------------------------------------
 
 {- | Spawn a process; return the exit value, throw on signal.  The `finally`
@@ -146,10 +167,9 @@ doProc ∷ ∀ ε ζ ω σ μ .
           ToMaybeTexts ω, MakeProc ζ, OutputHandles ζ ω,
           AsProcExitError ε, AsCreateProcError ε, AsFPathError ε,
           AsIOError ε, Printable ε, MonadError ε μ) ⇒
-         μ () → ExceptT ε μ σ → AbsFile → [𝕋] → μ (Word8, ω)
-doProc finally input exec as = do
-  let cmd = anyExit (mkCmd exec as)
-  result ← ѥ $ input ≫ \ l → system l cmd
+         μ () → σ → CmdSpec → μ (Word8, ω)
+doProc finally input cspec = do
+  result ← ѥ $ systemx input cspec
   finally
   case result of
     𝕷 e      → throwError e
@@ -157,6 +177,6 @@ doProc finally input exec as = do
       case ex of
         ExitVal x → return (x,w)
         -- we throw if the subprocess sees any signal
-        ExitSig _ → throwError (asProcExitError (cmd ⊣ cmdSpec) ex (𝕹,𝕹))
+        ExitSig _ → throwError (asProcExitError cspec ex (𝕹,𝕹))
 
 -- that's all, folks! ----------------------------------------------------------